01/requirements.txt
beautifulsoup4
Flask==3.1.0
pyspellchecker
requests==2.32.3